Working envelope verify signature
